+++ CHANGELOG.md
@@ -10,6 +10,7 @@ changed functionality, fixed bugs).
   to edit `flamenco-worker.cfg`.
 - Only fail UPnP/SSDP discovery when it fails to send on both IPv4 and IPv6.
 - Creating distribution files using [PyInstaller](http://www.pyinstaller.org/).
+- Fixed UPnP/SSDP discovery issues on Windows.
 
 
 ## Version 2.0.7 (released 2017-07-04)